# infra/openapi.json
{"openapi":"3.1.0","info":{"title":"Heizungsleser V2","description":"Backend für Multi-Mandanten Heizungsdaten-Visualisierung","version":"0.1.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login Access Token","description":"OAuth2 compatible token login, get an access token for future requests.","operationId":"login_access_token_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_access_token_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Read User Me","description":"Get current user.","operationId":"read_user_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/tenants/":{"get":{"tags":["tenants"],"summary":"Read Tenants","description":"Retrieve tenants. (Admin only)","operationId":"read_tenants_api_v1_tenants__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tenant"},"title":"Response Read Tenants Api V1 Tenants  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["tenants"],"summary":"Create Tenant","description":"Create new tenant. (Admin only)","operationId":"create_tenant_api_v1_tenants__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tenants/{tenant_id}":{"get":{"tags":["tenants"],"summary":"Read Tenant","description":"Get tenant by ID.","operationId":"read_tenant_api_v1_tenants__tenant_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["tenants"],"summary":"Update Tenant","description":"Update a tenant. (Admin only)","operationId":"update_tenant_api_v1_tenants__tenant_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/devices/":{"get":{"tags":["devices"],"summary":"Read Devices","description":"Retrieve devices for a tenant.","operationId":"read_devices_api_v1_devices__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"tenant_id","in":"query","required":true,"schema":{"type":"integer","title":"Tenant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Device"},"title":"Response Read Devices Api V1 Devices  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["devices"],"summary":"Create Device","description":"Create new device.","operationId":"create_device_api_v1_devices__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/devices/{device_id}":{"get":{"tags":["devices"],"summary":"Read Device","description":"Get device by ID.","operationId":"read_device_api_v1_devices__device_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"type":"integer","title":"Device Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["devices"],"summary":"Update Device","description":"Update a device.","operationId":"update_device_api_v1_devices__device_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"type":"integer","title":"Device Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/data/{device_id}/entities":{"get":{"tags":["data"],"summary":"Read Device Entities","description":"List entities for a device by introspecting InfluxDB.","operationId":"read_device_entities_api_v1_data__device_id__entities_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"type":"integer","title":"Device Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Entity"},"title":"Response Read Device Entities Api V1 Data  Device Id  Entities Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/data/{device_id}/timeseries":{"get":{"tags":["data"],"summary":"Read Device Timeseries","description":"Get timeseries data for specific entities of a device.","operationId":"read_device_timeseries_api_v1_data__device_id__timeseries_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"type":"integer","title":"Device Id"}},{"name":"entity_ids","in":"query","required":true,"schema":{"type":"array","items":{"type":"string"},"title":"Entity Ids"}},{"name":"start","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"ISO-8601 start time","title":"Start"},"description":"ISO-8601 start time"},{"name":"end","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"ISO-8601 end time","title":"End"},"description":"ISO-8601 end time"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDataResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_access_token_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_access_token_api_v1_auth_login_post"},"DataPoint":{"properties":{"ts":{"type":"string","title":"Ts"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"}},"type":"object","required":["ts"],"title":"DataPoint"},"Device":{"properties":{"display_name":{"type":"string","title":"Display Name"},"is_active":{"type":"boolean","title":"Is Active","default":true},"source_type":{"type":"string","title":"Source Type","default":"influxdb_v3"},"influx_database_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Influx Database Name"},"retention_policy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retention Policy"},"source_config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Source Config"},"id":{"type":"integer","title":"Id"},"tenant_id":{"type":"integer","title":"Tenant Id"},"slug":{"type":"string","title":"Slug"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["display_name","id","tenant_id","slug","created_at","updated_at"],"title":"Device"},"DeviceCreate":{"properties":{"display_name":{"type":"string","title":"Display Name"},"is_active":{"type":"boolean","title":"Is Active","default":true},"source_type":{"type":"string","title":"Source Type","default":"influxdb_v3"},"influx_database_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Influx Database Name"},"retention_policy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retention Policy"},"source_config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Source Config"},"tenant_id":{"type":"integer","title":"Tenant Id"}},"type":"object","required":["display_name","tenant_id"],"title":"DeviceCreate"},"DeviceDataResponse":{"properties":{"device_id":{"type":"integer","title":"Device Id"},"range":{"type":"object","title":"Range"},"series":{"items":{"$ref":"#/components/schemas/TimeSeriesResponse"},"type":"array","title":"Series"}},"type":"object","required":["device_id","range","series"],"title":"DeviceDataResponse"},"DeviceUpdate":{"properties":{"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"influx_database_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Influx Database Name"},"retention_policy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retention Policy"},"source_config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Source Config"}},"type":"object","title":"DeviceUpdate"},"Entity":{"properties":{"entity_id":{"type":"string","title":"Entity Id"},"domain":{"type":"string","title":"Domain"},"friendly_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Friendly Name"},"data_kind":{"type":"string","title":"Data Kind"},"chartable":{"type":"boolean","title":"Chartable","default":false},"icon":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icon"},"device_class":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Device Class"},"options":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Options"},"last_seen":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Seen"},"source_table":{"type":"string","title":"Source Table"}},"type":"object","required":["entity_id","domain","data_kind","source_table"],"title":"Entity"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Tenant":{"properties":{"name":{"type":"string","title":"Name"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"slug":{"type":"string","title":"Slug"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","id","slug","created_at","updated_at"],"title":"Tenant"},"TenantCreate":{"properties":{"name":{"type":"string","title":"Name"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["name"],"title":"TenantCreate"},"TenantUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"TenantUpdate"},"TimeSeriesResponse":{"properties":{"entity_id":{"type":"string","title":"Entity Id"},"friendly_name":{"type":"string","title":"Friendly Name"},"domain":{"type":"string","title":"Domain"},"data_kind":{"type":"string","title":"Data Kind"},"chartable":{"type":"boolean","title":"Chartable"},"points":{"items":{"$ref":"#/components/schemas/DataPoint"},"type":"array","title":"Points"},"meta":{"type":"object","title":"Meta"}},"type":"object","required":["entity_id","friendly_name","domain","data_kind","chartable","points","meta"],"title":"TimeSeriesResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"is_superuser":{"type":"boolean","title":"Is Superuser"}},"type":"object","required":["id","is_superuser"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}